draw the board different
I feel like making word the size of the board will make collision detection easier. This way makes the screen size the board size by changing the width and height the side of the block.

We don't have to keep it this way but i feel like drawing it this way will make the collision detection and edge detection easier.